Use a new pass split_loads for -mlate-explicit-relocs.  Move it tentatively before combine.
We get a bit of accidental scheduling by running this before lra.
In theory, running the splitter before combine should also allow other
opportunities.
The time the pass runs at can be moved relatively freely, both before
and after regalloc.
To make the most of the splitting, we should use a pipeline model that
matches the used / simulated core.